+++ src/user/dtos/login_2fa.go
@@ -1,0 +1,5 @@
+package user_dtos
+
+type Login2FARequest struct {
+	Code string `example:"random2FACode"`
+}