+++ src/utils/usecases/code.go
@@ -1,10 +1,22 @@
 package utils_usecases
 
 import (
+	cryptorand "crypto/rand"
+	"encoding/base32"
 	"math/rand"
 	"strconv"
 	"time"
 )
+
+func GenerateRandomSecret(length int) (string, error) {
+	randomBytes := make([]byte, length)
+	_, err := cryptorand.Read(randomBytes)
+	if err != nil {
+		return "", err
+	}
+	secret := base32.StdEncoding.EncodeToString(randomBytes)
+	return secret, nil
+}
 
 func GenerateRandomCode() string {
 	source := rand.NewSource(time.Now().UnixNano())